+++ docs/form_qc_coordinator/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## 1.1.0
 * Adds support for standalone forms submission
+* Strips leading zeros from PTID in error log name
   
 ## 1.0.1
 * Upgrades to dependencies