+++ docs/form_qc_coordinator/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 All notable changes to this gear are documented in this file.
 
-## Unreleased
-
+## 1.2.0
+* Adds functionality for handling multiple pipelines (submission, finalization)
 * Updates to read in files with `utf-8-sig` to handle BOM encoding
 
 ## 1.1.0